fix: made the StageTrampCommand be better
